Improved / Fixed JSON Writing
No longer hoards resources when writing and implemented new "new lap" detection. This works much better, however it will for the time being record and save a Lap 0 (Outlap). This can just be ignored.

The issue of having many entries from the next lap saved also persists, however, instead of having many JSON entries it is only one - this is an improvement but still needs to be fixed.