Fixed ValidLap Detection
For the first second of the session loading, the game assumes that when in the pits, the car has a positive track distance. As such, this set the valid lap to true for a split second and caused the JSON file to have unwanted data added to it.

This has been fixed by checking the session time and lap distance.